requirements.txt
# Data processing
# Note: fabric-sdk is pre-installed in the Fabric container environment
numpy>=1.24.0
pandas>=2.0.0

# API requests
requests>=2.31.0
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Typing and validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging
structlog>=23.1.0

# Date/time utilities
python-dateutil>=2.8.2
